requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
emoji==0.5.4
ephem==3.7.7.0
future==0.17.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
pyephem==3.7.7.0
pylint==2.3.1
python-telegram-bot==12.0.0
requests==2.22.0
six==1.12.0
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
